destructure redux-form-fields
